+++ application/views/modules/child/health/food.php
@@ -45,8 +45,10 @@
                     <?php echo $item->comment; ?>
                 </td>
                 <td>
+            <?php if(!is('parent')): ?>
                     <a class="delete" href="<?php echo site_url(); ?>child/deleteFoodPref/<?php echo $item->id; ?>">
                         <span class="fa fa-trash-o cursor"></span>
+                        <?php endif; ?>
                 </td>
             </tr>
             <?php
